rds-proxy-lambda/lambda/requirements.txt
# Lambda layer requirements
# For production, build a Lambda Layer with these dependencies

psycopg2-binary>=2.9.0
